refactor api service
